backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==1.55.0
python-dotenv==1.0.1
pydantic==2.10.0
pypdf==5.1.0
httpx<0.28.0
